Refactor auth
This is rather huge and I wish it was split more commits but most of the
code didnt compile through this process.

This makes AuthParms and AuthOptions direct arguments to the auth
functions, doing away with the builder functions. Having a separate
builder type for both functions, and each having functions to set all
the auth options/parms options seemed quite a bit overkill. Especially
as the user may wish to save these structs for reuse. I also found it
a bit confusing personally.

I've also tweaked what is and isn't an option. This is surely not 100%
correct and the spec is not the best at laying this out so I'm sure more
work will be needed.

This also combines things that can't be set at the same time into enums.
For auth at anyway, this still needs to be done for client options.