Implement ability to get existing channels and set their options
While there was already the ability to create channels via the get
method, channels were never actually stored anywhere and recalled again
when get was called a second time as the spec demands.

This introduces mutability into the REST SDK for the first time. The way
I've choosen to implement it here is to have a Vec of channels inside
the rest struct protected by a mutex. The channels themselves are actually
just names and look up the mutex'd list any time the options are
needed. This allows the settings to be upated at any time and the
updates to apply to existing channels.

This method does mean that the channel list needs to be locked on every
options access. This could be fixed by making each options instance it's
own mutex and arc but I don't think that's needed. Especially as there's
no way for a user to hold on to a mutex lock.

There's also the question of what happens if you release a channel then
try and publish/get options of that channel. The spec doesn't say
anything about this so for now it just uses the default options.

I've also removed the channel builder for now as there's only
2 arguments for a channel, maybe even one if we remove the with_options
function as the spec is unclear on if it should still be implemented.